chore: update versioning to v0.0.1-canary and RIFT v1.2
- Created scripts/set-version.sh for automated version management.
- Updated VERSION, Cargo.toml, build.gradle.kts, package.json, and documentation.

+++ crates/wavry-platform/src/dummy.rs
@@ -35,7 +35,7 @@ impl InputInjector for DummyInjector {
         Ok(())
     }
 
-    fn mouse_absolute(&mut self, x: i32, y: i32) -> Result<()> {
+    fn mouse_absolute(&mut self, x: f32, y: f32) -> Result<()> {
         info!("DummyInjector: Mouse Absolute {}, {}", x, y);
         Ok(())
     }
